requirements.txt
# Core
numpy>=1.24
pandas>=1.5

# API / concurrency (only if you use FastAPI)
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.0

# HTTP (for Ollama)
requests>=2.31.0

# Logging / utils
loguru>=0.7
tqdm>=4.66

# RAG (FAISS-based)
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pymupdf>=1.23.0          # PDF reading (fitz)


# Optional (for experiments / evaluation)
scikit-learn>=1.3
matplotlib>=3.7
pytest>=7.0